pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "inngest>=0.5.13",
    "llama-index-core>=0.14.8",
    "llama-index-readers-file>=0.5.5",
    "openai>=2.8.1",
    "pandas>=2.2.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.0.1",
    "python-env>=1.0.0",
    "qdrant-client>=1.16.1",
    "scikit-learn>=1.7.2",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
